requirements.txt
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.5.0
PyQt6-Fluent-Widgets>=1.4.0
PyQt6-Frameless-Window>=0.3.0
pywin32>=306
psutil>=5.9.0
pyautogui>=0.9.54
opencv-python>=4.8.0
requests>=2.31.0
cloudscraper>=1.2.71
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.11.0
